requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlmodel==0.0.22
psycopg2-binary==2.9.9
boto3==1.29.3
python-multipart==0.0.6
pgvector==0.3.6
markdown==3.7
pydantic==2.10.2
python-dotenv==1.0.1
nltk==3.9.1
transformers[torch]==4.35.2